+++ src/app/components/tree/base/terra-base-tree.component.ts
@@ -3,7 +3,10 @@ import {
     OnInit
 } from '@angular/core';
 import { TerraLeafInterface } from '../leaf/terra-leaf.interface';
-import { isNull } from 'util';
+import {
+    isNull,
+    isNullOrUndefined
+} from 'util';
 
 /**
  * @author mkunze
@@ -56,15 +59,14 @@ export class TerraBaseTreeComponent implements OnInit
         }
     }
 
-
-    private onLeafClick(clickedLeaf:TerraLeafInterface):void
+    protected onLeafClick(clickedLeaf:TerraLeafInterface):void
     {
-        if(!isNull(clickedLeaf.subLeafList) && !clickedLeaf.avoidOpenOnClick)
+        if(!isNullOrUndefined(clickedLeaf.subLeafList) && !clickedLeaf.avoidOpenOnClick)
         {
             this.toggleOpen(clickedLeaf);
         }
 
-        if(!isNull(clickedLeaf.clickFunction) && !clickedLeaf.isActive)
+        if(!isNullOrUndefined(clickedLeaf.clickFunction) && !clickedLeaf.isActive)
         {
             clickedLeaf.clickFunction();
         }
@@ -106,17 +108,6 @@ export class TerraBaseTreeComponent implements OnInit
         clickedLeaf.isOpen = !clickedLeaf.isOpen;
     }
 
-    private onArrowClick(clickedLeaf:TerraLeafInterface):void
-    {
-        if(!isNull(clickedLeaf.onOpenFunction) && !clickedLeaf.isOnOpenFunctionCalled)
-        {
-            clickedLeaf.onOpenFunction();
-            clickedLeaf.isOnOpenFunctionCalled = true;
-        }
-
-        this.toggleOpen(clickedLeaf);
-    }
-
     private recursiveSearchActiveLeaf(leafListToSearch:Array<TerraLeafInterface>):TerraLeafInterface
     {
         let foundLeaf:TerraLeafInterface;
@@ -133,7 +124,7 @@ export class TerraBaseTreeComponent implements OnInit
             {
                 foundLeaf = this.recursiveSearchActiveLeaf(leaf.subLeafList);
 
-                if(!isNull(foundLeaf))
+                if(!isNullOrUndefined(foundLeaf))
                 {
                     break;
                 }
@@ -146,10 +137,5 @@ export class TerraBaseTreeComponent implements OnInit
     public getSelectedLeaf():TerraLeafInterface
     {
         return this.recursiveSearchActiveLeaf(this.inputLeafList);
-    }
-
-    private checkIfArrowNeeded(clickedLeaf:TerraLeafInterface):boolean
-    {
-        return !isNull(clickedLeaf.subLeafList) || !isNull(clickedLeaf.onOpenFunction);
     }
 }