+++ CHANGELOG.md
@@ -6,12 +6,17 @@
 
 ### Bug Fixes
 * **terra-node-tree** fixed issues occuring when adding a new node whilst an existing node had been selected
+* **terra-info** fix export of component
 
 <a name="3.8.0"></a>
 # 3.8.0 (09.04.2019)
 
 ### Features
 * **terra-form** set empty array as default value for list controls.
+* **terra-checkbox** temporarily added `notifyOnChanges` to be notified on leaf change
+* **terra-checkbox-tree** 
+	- CSS changes
+	- added functionality for disabled leaf
 
 <a name="3.7.0"></a>
 # 3.7.0 (27.03.2019)