+++ CHANGELOG.md
@@ -1,9 +1,22 @@
+<a name="3.9.0"></a>
+# 3.9.0 (Unreleased)
+
+### Bug Fixes
+* **terra-node-tree** fixed issues occuring when adding a new node whilst an existing node had been selected
+* **terra-info** fix export of component
+
+### Features
+* **tc-radio-group** now uses a generated name if it is not set via input.
+* **terra-checkbox-tree** 
+	- CSS changes
+	- added functionality for disabled leaf
+* **terra-checkbox** temporarily added `notifyOnChanges` to be notified on leaf change
+
 <a name="3.8.0"></a>
-# 3.8.0 (NOT RELEASED)
+# 3.8.0 (09.04.2019)
 
 ### Features
 * **terra-form** set empty array as default value for list controls.
-* **tc-radio-group** now uses a generated name if it is not set via input.
 
 <a name="3.7.0"></a>
 # 3.7.0 (27.03.2019)